% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{clean}
\alias{clean}
\alias{clean.messages}
\title{Clean Tweet}
\usage{
clean(x, ...)

\method{clean}{messages}(x, log = NULL, keep = 1:10, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of class \code{messages}.}

\item{...}{Currently not used.}

\item{log}{A \code{\link{character}} string giving the path of the log file
(see below).}

\item{keep}{An \code{\link{integer}} vector giving the indices of the messages to be
kept.}
}
\value{
A \code{\link{list}} of \code{\link{character}} string (tweet message).
}
\description{
Clean Tweet
}
\section{Prevent Double Post}{

A log file can be used to avoid posting the same documents several times.
This file records the unique identifier of each message (generated by
\code{compose()}). The log file must have four columns separated by a single
white space:
\itemize{
\item Date (YYYY-MM-DD);
\item Time (HH-MM-SS);
\item Tweet status id;
\item Message unique id.
}
}

\author{
N. Frerebeau
}
\keyword{internal}
