% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{compose}
\alias{compose}
\alias{composeTweet}
\alias{composeTweet.hal_document}
\title{Compose Tweet}
\usage{
composeTweet(x, ...)

\method{composeTweet}{hal_document}(x, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of document references.}

\item{...}{Currently not used.}
}
\value{
Returns a \code{\link{character}} string (tweet message).
}
\description{
Compose Tweet
}
\author{
N. Frerebeau
}
