% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{get_hal}
\alias{get_hal}
\alias{get_hal_team}
\alias{get_hal_author}
\title{Get Documents From HAL}
\usage{
get_hal_team(id, limit = 10)

get_hal_author(id, limit = 10)
}
\arguments{
\item{id}{A \code{\link{character}} string giving the structure or the author (IdHal)
id.}

\item{limit}{An \code{\link{integer}} specifying the number of results to be returned.}

\item{type}{A \code{\link{character}} string giving the type of the documents to be
returned (see HAL API documentation).}
}
\value{
A \code{\link{list}} of document references. Each element of the \code{list} is a \code{list} of
class \code{hal_document}.
}
\description{
Get Documents From HAL
}
\details{
Retrieves documents from HAL published during the last year.
}
\examples{
\dontrun{
get_hal_team("399901")
}
}
\author{
N. Frerebeau
}
