% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{get_rss}
\alias{get_rss}
\title{Get Posts From an RSS Feed}
\usage{
get_rss(feed, limit = 7 * 24 * 60 * 60)
}
\arguments{
\item{feed}{A \code{\link{character}} string giving the url for the feed that you want
to get.}

\item{limit}{An \code{\link{integer}} specifying the time limit (in second) back from
today (defaults to \eqn{604800}, i.e. one week).}
}
\description{
Get Posts From an RSS Feed
}
\examples{
\dontrun{
get_rss("https://www.archeosciences-bordeaux.fr/spip.php?page=backend-breves")
}
}
\author{
N. Frerebeau
}
