% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{hal}
\alias{hal}
\alias{get_hal_team}
\title{Get Documents From HAL}
\usage{
get_hal_team(id, limit = 10)
}
\arguments{
\item{id}{A \code{\link{character}} string giving the structure or the author id.}

\item{limit}{An \code{\link{integer}} specifying the number of results to be returned.}
}
\value{
A \code{\link{list}} of document references. Each element of the \code{list} is a \code{list} of
class \code{hal_document}.
}
\description{
Get Documents From HAL
}
\examples{
\dontrun{
get_hal_team("399901")
}
}
\author{
N. Frerebeau
}
