% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logger}
\alias{logger}
\title{Log}
\usage{
logger(f, file)
}
\arguments{
\item{f}{A \code{\link{function}}.}

\item{file}{A \code{\link{character}} string naming the file to print the output
of \code{f}.}
}
\value{
\code{logger()} is a decorator function called for its side-effect. Returns
the output of \code{f}.
}
\description{
Logs the output of any function.
}
\author{
N. Frerebeau
}
\keyword{internal}
