% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{send}
\alias{send}
\alias{sendTweet}
\alias{sendTweet.hal_document}
\title{Send Tweet}
\usage{
sendTweet(x, ...)

\method{sendTweet}{hal_document}(x, log, silent = TRUE, test = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of document references.}

\item{...}{Currently not used.}

\item{log}{A \code{\link{character}} string giving the path of the log file.}

\item{silent}{A \code{\link{logical}} scalar: should the report of error messages be
suppressed?}

\item{test}{A \code{\link{logical}} scalar.}
}
\value{
Returns a \code{\link{logical}} scalar.
}
\description{
Send Tweet
}
\author{
N. Frerebeau
}
