% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tweet.R
\name{tweet}
\alias{tweet}
\alias{composeTweet}
\alias{sendTweet}
\alias{composeTweet.hal_document}
\alias{sendTweet.hal_document}
\title{Compose Tweet}
\usage{
composeTweet(x, ...)

sendTweet(x, ...)

\method{composeTweet}{hal_document}(x, ...)

\method{sendTweet}{hal_document}(x, log, silent = TRUE, test = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of document references.}

\item{...}{Currently not used.}

\item{log}{A \code{\link{character}} string giving the path of the log file.}

\item{silent}{A \code{\link{logical}} scalar: should the report
of error messages be suppressed?}

\item{test}{A \code{\link{logical}} scalar.}
}
\value{
\code{composeTweet} returns a \code{\link{list}} of character strings
 (tweet messages).

 \code{sendTweet} returns a \code{\link{logical}} scalar.
}
\description{
Compose Tweet
}
\author{
N. Frerebeau
}
