% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterbot-package.R
\docType{package}
\name{twitterbot-package}
\alias{twitterbot-package}
\alias{_PACKAGE}
\alias{twitterbot}
\title{twitterbot: A Twitter Bot for the Lab}
\description{
Accesses documents from the Open
    Archive HAL <https://hal.archives-ouvertes.fr> and tweets the lastest
    scientific publications.
}
\details{
\tabular{ll}{
\strong{Package:} \tab twitterbot \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.1 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nfrerebeau/twitterbot}
  \item Report bugs at \url{https://github.com/nfrerebeau/twitterbot/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

*\emph{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
