% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{clean}
\alias{clean}
\alias{clean.messages}
\title{Clean Tweet}
\usage{
clean(x, ...)

\method{clean}{messages}(x, log = NULL, keep = 1:10, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of class \code{messages}.}

\item{...}{Currently not used.}

\item{log}{A \code{\link{character}} string giving the path of the log file
(see below).}

\item{keep}{An \code{\link{integer}} vector giving the indices of the messages to be
kept.}
}
\value{
A \code{\link{list}} of \code{\link{character}} string (tweet message).
}
\description{
Clean Tweet
}
\author{
N. Frerebeau
}
