% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{compose}
\alias{compose}
\alias{compose.hal_documents}
\alias{compose.hal_document}
\alias{compose.rss_items}
\alias{compose.rss_item}
\title{Compose Tweet}
\usage{
compose(x, ...)

\method{compose}{hal_documents}(x, ...)

\method{compose}{hal_document}(x, ...)

\method{compose}{rss_items}(x, ...)

\method{compose}{rss_item}(x, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of document references
(typically a list of class \code{\link[=get_hal_team]{hal_documents}}).}

\item{...}{Currently not used.}
}
\value{
Returns a \code{\link{character}} string (tweet message) or a \code{\link{list}} of \code{character}
strings.
}
\description{
Compose Tweet
}
\author{
N. Frerebeau
}
