% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/hal.R
\name{hal}
\alias{hal}
\alias{getDocuments}
\alias{cleanDocuments}
\title{Get Documents From HAL}
\usage{
getDocuments(id, limit = 10)

cleanDocuments(x, log, keep = 1:10)
}
\arguments{
\item{id}{A \code{\link{character}} string giving the structure id.}

\item{limit}{An \code{\link{integer}} specifying the number of results to be
returned.}

\item{x}{A \code{\link{list}} of document references.}

\item{log}{A \code{\link{data.frame}}.}

\item{keep}{An \code{\link{integer}} vector giving the indices of the
documents to be kept.}
}
\value{
A \code{\link{list}} of document references. Each element of the list
 is a list of class \code{hal_document}.
}
\description{
Get Documents From HAL
}
\author{
N. Frerebeau
}
