% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tweet.R
\name{log}
\alias{log}
\alias{readLog}
\alias{writeLog}
\title{Log}
\usage{
readLog(file, silent = TRUE)

writeLog(file, tweet_id, doc_id)
}
\arguments{
\item{file}{A \code{\link{character}} string giving the name of the file
which the data are to be read from. Each row of the table appears as one
line of the file (see \code{\link[utils]{read.table}}).}

\item{silent}{A \code{\link{logical}} scalar: should the report
of error messages be suppressed?}

\item{tweet_id}{A \code{\link{character}} string giving the id of the tweet.}

\item{doc_id}{A \code{\link{character}} string giving the id of the HAL
document.}

\item{...}{Extra parameters to be passed to \code{\link[utils]{read.table}}.}
}
\value{
\code{readLog} returns a \code{\link{data.frame}} if the log file exists and
 contains at least one row, else returns \code{FALSE}.

 \code{writeLog} returns \code{NULL}.
}
\description{
Log
}
\details{
The log file must have four columns separated by a single white space:
 \itemize{
  \item{Date (YYYY-MM-DD).}
  \item{Time (HH-MM-SS).}
  \item{Tweet id.}
  \item{HAL id.}
 }
}
\seealso{
\link{cat}
}
\author{
N. Frerebeau
}
