% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{post}
\alias{post}
\alias{post.hal_documents}
\alias{post.messages}
\alias{post.message}
\title{Post Tweet}
\usage{
post(x, ...)

\method{post}{hal_documents}(x, log = NULL, keep = 1:10, silent = TRUE, test = FALSE, ...)

\method{post}{messages}(x, silent = TRUE, test = FALSE, ...)

\method{post}{message}(x, silent = TRUE, test = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of messages.}

\item{...}{Currently not used.}

\item{log}{A \code{\link{character}} string giving the path of the log file (see
below).}

\item{keep}{An \code{\link{integer}} vector giving the indices of the messages to be
kept.}

\item{silent}{A \code{\link{logical}} scalar: should the report of error messages be
suppressed?}

\item{test}{A \code{\link{logical}} scalar.}
}
\value{
Invisibly returns the tweet date, time and status id.
}
\description{
Post Tweet
}
\section{Prevent Double Post}{

A log file can be used to avoid posting the same documents several times.
This file records the unique identifier of each message (generated by
\code{compose()}). The log file must have four columns separated by a single
white space:
\itemize{
\item Date (YYYY-MM-DD);
\item Time (HH-MM-SS);
\item Tweet status id;
\item Message unique id.
}
}

\author{
N. Frerebeau
}
