% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{post}
\alias{post}
\alias{post.hal_documents}
\alias{post.messages}
\alias{post.message}
\title{Post Tweet}
\usage{
post(x, ...)

\method{post}{hal_documents}(x, log = NULL, keep = 1:10, silent = TRUE, test = FALSE, ...)

\method{post}{messages}(x, silent = TRUE, test = FALSE, ...)

\method{post}{message}(x, silent = TRUE, test = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{list}} of messages.}

\item{...}{Currently not used.}

\item{log}{A \code{\link{character}} string giving the path of the log file.}

\item{keep}{An \code{\link{integer}} vector giving the indices of the messages to be
kept.}

\item{silent}{A \code{\link{logical}} scalar: should the report of error messages be
suppressed?}

\item{test}{A \code{\link{logical}} scalar.}
}
\value{
Returns a \code{\link{logical}} scalar.
}
\description{
Post Tweet
}
\author{
N. Frerebeau
}
