% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterbot-package.R
\docType{package}
\name{twitterbot-package}
\alias{twitterbot-package}
\alias{_PACKAGE}
\alias{twitterbot}
\title{twitterbot: Tweet Last Publications From The IRAMAT-CRP2A Laboratory}
\description{
This package accesses documents from the Open
    Archive HAL <https://hal.archives-ouvertes.fr/> and tweets the last
    scientific publications from the CRP2A lab on a daily basis.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab twitterbot \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.1 \cr
 \strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crp2a/twitterbot}
  \item Report bugs at \url{https://github.com/crp2a/twitterbot/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
